+++ source/App/source/Common.Tests/Extensions/DependencyInjection/IdentityExtensionsTests.cs
@@ -29,32 +29,90 @@ public class IdentityExtensionsTests
 
     private ServiceCollection Services { get; }
 
+    #region AddTokenCredentialProvider
+
     [Fact]
-    public void Given_Services_When_AddAuthorizationHeaderProvider_Then_RegistrationsArePerformed()
+    public void Given_Services_When_AddTokenCredentialProvider_Then_RegistrationsArePerformed()
     {
         // Act
-        Services.AddAuthorizationHeaderProvider();
+        Services.AddTokenCredentialProvider();
 
         // Assert
         var serviceProvider = Services.BuildServiceProvider();
 
-        var actualAuthorizationHeaderProvider = serviceProvider.GetRequiredService<IAuthorizationHeaderProvider>();
-        actualAuthorizationHeaderProvider.Should().NotBeNull();
+        var actual = serviceProvider.GetRequiredService<TokenCredentialProvider>();
+        actual.Should().NotBeNull();
     }
 
     [Fact]
-    public void Given_AddAuthorizationHeaderProviderWasCalled_When_AddAuthorizationHeaderProvider_Then_RegistrationsArePerformedOnlyOnce()
+    public void Given_AddTokenCredentialProviderWasCalled_When_AddTokenCredentialProvider_Then_RegistrationsArePerformedOnlyOnce()
+    {
+        // Arrange
+        Services.AddTokenCredentialProvider();
+
+        // Act
+        Services.AddTokenCredentialProvider();
+
+        // Assert
+        var serviceProvider = Services.BuildServiceProvider();
+
+        var actual = serviceProvider.GetServices<TokenCredentialProvider>();
+        actual.Count().Should().Be(1);
+    }
+
+    #endregion
+
+    #region AddAuthorizationHeaderProvider
+
+    [Fact]
+    public void Given_RequiredServicesNotRegisteredAndAddAuthorizationHeaderProvider_When_GetRequiredService_Then_ThrowsException()
     {
         // Arrange
         Services.AddAuthorizationHeaderProvider();
 
+        var serviceProvider = Services.BuildServiceProvider();
+
+        // Act
+        var act = () => serviceProvider.GetRequiredService<IAuthorizationHeaderProvider>();
+
+        // Assert
+        act.Should()
+            .Throw<InvalidOperationException>()
+            .WithMessage("No service for type 'Energinet.DataHub.Core.App.Common.Identity.TokenCredentialProvider' has been registered*");
+    }
+
+    [Fact]
+    public void Given_RequiredServices_When_AddAuthorizationHeaderProvider_Then_RegistrationsArePerformed()
+    {
+        // Arrange
+        Services.AddTokenCredentialProvider();
+
         // Act
         Services.AddAuthorizationHeaderProvider();
 
         // Assert
         var serviceProvider = Services.BuildServiceProvider();
 
-        var actualAuthorizationHeaderProviders = serviceProvider.GetServices<IAuthorizationHeaderProvider>();
-        actualAuthorizationHeaderProviders.Count().Should().Be(1);
+        var actual = serviceProvider.GetRequiredService<IAuthorizationHeaderProvider>();
+        actual.Should().NotBeNull();
     }
+
+    [Fact]
+    public void Given_RequiredServicesAndAddAuthorizationHeaderProviderWasCalled_When_AddAuthorizationHeaderProvider_Then_RegistrationsArePerformedOnlyOnce()
+    {
+        // Arrange
+        Services.AddTokenCredentialProvider();
+        Services.AddAuthorizationHeaderProvider();
+
+        // Act
+        Services.AddAuthorizationHeaderProvider();
+
+        // Assert
+        var serviceProvider = Services.BuildServiceProvider();
+
+        var actual = serviceProvider.GetServices<IAuthorizationHeaderProvider>();
+        actual.Count().Should().Be(1);
+    }
+
+    #endregion
 }