+++ source/App/source/Common.Tests/Identity/AuthorizationHeaderProviderTests.cs
@@ -41,20 +41,4 @@ public class AuthorizationHeaderProviderTests
         var token = tokenhandler.ReadJwtToken(actual.Parameter);
         token.Audiences.Should().Contain(SubsystemAuthenticationOptionsForTests.ApplicationIdUri);
     }
-
-    [Fact(Skip = "This test is not consistently true, because the test might run at a time where the token is refreshed on the CI environment.")]
-    public async Task Given_ReusedSutInstance_When_CreateAuthorizationHeaderIsCalledMultipleTimes_Then_ReturnedHeaderContainsSameTokenBecauseTokenCacheIsUsed()
-    {
-        // Arrange
-        var credential = new DefaultAzureCredential();
-        var sut = new AuthorizationHeaderProvider(credential);
-
-        // Act
-        var header01 = sut.CreateAuthorizationHeader(SubsystemAuthenticationOptionsForTests.ApplicationIdUri);
-        await Task.Delay(TimeSpan.FromSeconds(1));
-        var header02 = sut.CreateAuthorizationHeader(SubsystemAuthenticationOptionsForTests.ApplicationIdUri);
-
-        // Assert
-        header01.Parameter.Should().Be(header02.Parameter);
-    }
 }